refactor(mobile): split leaf modules out of db/queries.ts
Conservative continuation of the queries.ts split (the coupled note/card/tag/
conflict core stays together — a full domain split would risk import cycles).
Extract the clean, low-coupling pieces:

- db/types.ts — the base row interfaces (NoteRow/ProjectRow/ColumnRow/CardRow/
  TagRow) + knowledge-graph types. Pure types, imported everywhere, zero cycle
  risk.
- db/row-helpers.ts — the pure parseIds() tag-id parser (used across domains).
- db/graph-queries.ts — getKnowledgeGraph(), depending only on getDb/sql/
  row-helpers/types (inlines its tag lookup instead of importing tagsForIds so
  it never imports back into queries.ts).
- db/breakout.ts — listBreakoutBricks() easter-egg sampler (getDb-only leaf).

queries.ts re-exports all three via export * so every @/db/queries import is
unchanged; it now imports parseIds + base types. queries.ts 1349 -> 1143.
No import cycles (verified: none of the new modules import ./queries).

Verified: mobile tsc + eslint . clean; getKnowledgeGraph body diff-identical
except the tagsForIds->inline-SQL swap (same query); mobile vitest (26) green.
