+++ mobile/changelogs/v0.1.3.md
@@ -66,6 +66,7 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - Split the Chat screen — which had grown to one 800-line file mixing message rendering, the composer, keyboard/scroll choreography and tool-navigation logic — into focused pieces (a message bubble, reasoning disclosure, tool-chip trail, composer, a keyboard-scroll hook, and shared/unit-tested tool-navigation logic), and made it load its saved history only once on open instead of twice. No visible change; the chat behaves and looks exactly the same, but the code is far easier to change safely.
 - Split the Calendar view (previously a single ~930-line file) into focused pieces — the month/week grid math, the day cell, task chips, the unscheduled tray, and the day-detail row — with the grid/date logic now covered by unit tests. Mobile logic tests are also wired into the shared test runner for the first time (the calendar date math and the drag auto-scroll helpers now run in CI). No visible change; the calendar looks and behaves exactly the same.
 - Moved the on-device semantic-search index queries (note + task embeddings) out of the 1,600-line database-queries file into their own module, and factored the shared SQL fragments into one place. Purely internal reorganisation — every existing import keeps working via a re-export; no behaviour change.
+- Continued splitting the database-queries file: the shared row/graph types, the knowledge-graph builder, the tag-id parser, and the breakout easter-egg sampler now live in their own small modules (re-exported so imports are unchanged). Purely internal, cycle-free reorganisation; no behaviour change.
 - Split the AI-settings form into smaller pieces (the provider/reasoning segmented buttons, the labelled input field, and the Private-Cloud-Compute usage bar now live in their own files, sharing one style module). Purely internal; the settings screen looks and behaves exactly the same.
 - Split the project detail screen — its note-row, folder-row, tab segment, filter bar and list-row helpers now live in their own files. Purely internal; the project's Overview / Notes / Board tabs look and behave exactly the same.
 - Moved the Knowledge Graph's WebView document builder (the bundled D3 library plus the force-simulation and radial-sunburst rendering code) out of the 1,000-line graph component into its own file. Purely internal; the graph renders and behaves exactly the same.
